Refactor day9, int64 -> int

+++ day09/day09_test.go
@@ -3,7 +3,7 @@ package day09
 import "testing"
 
 func TestRegression(t *testing.T) {
-	right := []int64{1504371145, 183278487}
+	right := []int{1504371145, 183278487}
 	left := []interface{}{RunPart1().Value, RunPart2().Value}
 
 	for i, l := range left {
